// src/api/generated/models/LanguageWithRegions.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Babeonym API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { LanguageRegion } from './LanguageRegion';
import {
    LanguageRegionFromJSON,
    LanguageRegionFromJSONTyped,
    LanguageRegionToJSON,
    LanguageRegionToJSONTyped,
} from './LanguageRegion';

/**
 * 
 * @export
 * @interface LanguageWithRegions
 */
export interface LanguageWithRegions {
    /**
     * 
     * @type {number}
     * @memberof LanguageWithRegions
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof LanguageWithRegions
     */
    label: string;
    /**
     * 
     * @type {Array<LanguageRegion>}
     * @memberof LanguageWithRegions
     */
    regions: Array<LanguageRegion>;
}

/**
 * Check if a given object implements the LanguageWithRegions interface.
 */
export function instanceOfLanguageWithRegions(value: object): value is LanguageWithRegions {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('label' in value) || value['label'] === undefined) return false;
    if (!('regions' in value) || value['regions'] === undefined) return false;
    return true;
}

export function LanguageWithRegionsFromJSON(json: any): LanguageWithRegions {
    return LanguageWithRegionsFromJSONTyped(json, false);
}

export function LanguageWithRegionsFromJSONTyped(json: any, ignoreDiscriminator: boolean): LanguageWithRegions {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'label': json['label'],
        'regions': ((json['regions'] as Array<any>).map(LanguageRegionFromJSON)),
    };
}

export function LanguageWithRegionsToJSON(json: any): LanguageWithRegions {
    return LanguageWithRegionsToJSONTyped(json, false);
}

export function LanguageWithRegionsToJSONTyped(value?: LanguageWithRegions | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'label': value['label'],
        'regions': ((value['regions'] as Array<any>).map(LanguageRegionToJSON)),
    };
}